Update minimum_viable_product.py
Added a match case block that sets subdivision type to either catmull clark or simple based on number id. Added a function that removes a modifier from a list of objects by name. that removes all modifiers. Added a new case that removes all modifiers from a list of objects.